URD: upload, resize and display

+++ src/main/java/eu/zavadil/ocr/storage/LocalPath.java
@@ -29,6 +29,7 @@ public interface LocalPath {
 	}
 
 	default void delete() {
+		if (!this.exists()) return;
 		try {
 			Files.delete(this.asPath());
 		} catch (IOException e) {
